requirements-docker.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database
pymysql==1.1.0
sqlalchemy==2.0.23
DBUtils==3.1.2

# Auth and encryption
PyJWT==2.8.0
bcrypt==4.1.1
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# LangChain and AI (pinned for fast install)
langchain==0.3.25
langchain-core==0.3.61
langchain-community==0.3.21
langchain-chroma==0.2.2
langchain-huggingface==0.1.2
langchain-ollama==0.3.3
sentence-transformers==3.4.1
chromadb==0.6.3
transformers==4.48.3
# PyTorch CPU-only (no CUDA, saves ~1.5GB)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.2
# 默書 forensic OCR (TrOCR + doctr line detection)
python-doctr[torch]>=0.8.0
openai==1.61.0

# HTTP requests
requests==2.31.0
httpx==0.28.1

# Image processing
Pillow==12.1.1
pillow-heif==1.2.1
qrcode[pil]==8.0

# Document processing
python-docx==1.1.0
python-pptx==1.0.2
PyPDF2==3.0.1
PyMuPDF==1.25.5
openpyxl==3.1.2

# NOTE: PyQt6 removed — desktop-only, not needed for server

# Data processing
pandas==2.2.3
numpy==1.26.4
scipy>=1.10.0

# Environment and config
python-dotenv==1.0.0
pydantic==2.12.5
pydantic-settings==2.13.1

# Utilities
markdown2==2.5.4
pygments==2.19.2
